+++ src/components/inputs/ArrayInput.vue
@@ -1,0 +1,28 @@
+<template>
+  <q-select
+    v-model="localValue"
+    use-input
+    use-chips
+    multiple
+    hide-dropdown-icon
+    input-debounce="0"
+    new-value-mode="add"
+    :rules="[
+      (value) => isRequired($t, value, attribute.definition?.required),
+    ]"
+  />
+</template>
+
+<script setup>
+import { ref } from 'vue';
+import { isRequired } from 'src/composables/QuasarFieldRule';
+
+const props = defineProps({
+  attribute: {
+    type: Object,
+    required: true,
+  },
+});
+
+const localValue = ref(props.attribute.value);
+</script>