+++ changelog.md
@@ -64,5 +64,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename GitForm and GitConfigurationDialog to GitSettingsForm and GitSettingsDialog respectively
 - Split git setting dialog in two different dialogs : GitAuthenticationDialog & GitAddRemoteDialog
 - Create a component from a template
+- Create ArrayInput
 
 [unreleased]: https://github.com/ditrit/leto-modelizer/blob/main/changelog.md#unreleased